requirements/base.txt
Brotli==1.1.0
dj-database-url==2.1.0
Django==4.2.11
django-environ==0.11.2
gunicorn==21.2.0
matplotlib==3.7.5
numpy==1.24.4
pandas==2.0.3
pillow==10.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
django-cloudinary-storage==0.3.0